Added hangman command, fixed issue with ls imports.
