+++ server-old.js
@@ -170,10 +170,10 @@
     //     var coin = require("./coinflip");
     //     message.channel.send(coin.flip(args, ifProfile(message.author.id), prefix, message));
     //   break;
-    case "dice":
-        var dice = require("./dice");
-        message.channel.send(dice.roll(args, ifProfile(message.author.id), prefix, message));
-      break;
+    // case "dice":
+    //     var dice = require("./dice");
+    //     message.channel.send(dice.roll(args, ifProfile(message.author.id), prefix, message));
+    //   break;
     // case "user:all":
     //   break;
     // case "about":
@@ -205,11 +205,11 @@
     /* case "guess":
         throw "CommandUtilizationError: This command does not exist yet!";
       break; */
-    case "hm":
-        //Hangman game commands
-        var hm = require("./hangman");
-        message.channel.send(hm.do(args, ifProfile(message.author.id), prefix, message));
-      break;
+    // case "hm":
+    //     //Hangman game commands
+    //     var hm = require("./hangman");
+    //     message.channel.send(hm.do(args, ifProfile(message.author.id), prefix, message));
+    //   break;
     // case "blackjack":
     //     var blackjack = require("./blackjack");
     //     if (ls.exist(message.author.id + "blackjackGame")) {
